MLE
Corrección de error al obtener detalle de ítem por categoría.

+++ app/src/main/java/cl/cdum/mlentertainment/data/model/categoryItemDetail/Struct.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Struct(
     @SerializedName("number")
-    val number: Int? = null,
+    val number: Double? = null,
     @SerializedName("unit")
     val unit: String? = null
 )
